Refactor UI of upload qr
