+++ core/ui/qr/components/styles.tsx
@@ -1,35 +1,29 @@
-import { PageContent } from '@lib/ui/page/PageContent'
 import styled from 'styled-components'
-
-export const Container = styled(PageContent)`
-  position: relative;
-  justify-content: space-between;
-`
 
 export const VideoWrapper = styled.div`
   align-self: center;
+  height: 368px;
   position: relative;
-  width: 365px;
-  height: 365px;
+  width: 368px;
 `
 
 export const BorderImageWrapper = styled.div`
-  position: absolute;
   inset: -1.2px;
+  position: absolute;
 
   & > img {
-    width: 100%;
     height: 100%;
     object-fit: cover;
+    width: 100%;
   }
 `
 
 export const Video = styled.video`
-  width: 100%;
+  background: rgba(4, 36, 54, 0.3);
+  border: 1px solid rgba(72, 121, 253, 0.15);
+  border-radius: 40px;
+  box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
   height: 100%;
   object-fit: cover;
-  border-radius: 40px;
-  border: 1px solid rgba(72, 121, 253, 0.15);
-  background: rgba(4, 36, 54, 0.3);
-  box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
+  width: 100%;
 `
