+++ core/ui/qr/hooks/useCameraPermissionQuery.tsx
