--- core/ui/qr/upload/UploadQrPage/index.tsx
@@ -1,23 +1,0 @@
-import { useCurrentVaultId } from '@core/ui/storage/currentVaultId'
-import { useVaults } from '@core/ui/storage/vaults'
-import { CurrentVaultProvider } from '@core/ui/vault/state/currentVault'
-import { getVaultId } from '@core/ui/vault/Vault'
-
-import { UploadQrPageWithExistingVault } from './UploadQrPageWithExistingVault'
-import { UploadQrPageWithoutVault } from './UploadQrPageWithoutVault'
-
-export const UploadQrPage = () => {
-  const currentVaultId = useCurrentVaultId()
-  const vaults = useVaults()
-  const vault = vaults.find(vault => getVaultId(vault) === currentVaultId)
-
-  if (!vault) {
-    return <UploadQrPageWithoutVault />
-  }
-
-  return (
-    <CurrentVaultProvider value={vault}>
-      <UploadQrPageWithExistingVault />
-    </CurrentVaultProvider>
-  )
-}
