+++ core/ui/qr/upload/index.tsx
@@ -1,0 +1,96 @@
+import { useCoreNavigate } from '@core/ui/navigation/hooks/useCoreNavigate'
+import { useCoreViewState } from '@core/ui/navigation/hooks/useCoreViewState'
+import { ScanQrView } from '@core/ui/qr/components/ScanQrView'
+import { UploadQrView } from '@core/ui/qr/components/UploadQrView'
+import { useDeriveChainFromWalletAddress } from '@core/ui/qr/hooks/useDeriveChainFromWalletAddress'
+import { useCurrentVaultId } from '@core/ui/storage/currentVaultId'
+import { useVaults } from '@core/ui/storage/vaults'
+import { getVaultId } from '@core/ui/vault/Vault'
+import { Match } from '@lib/ui/base/Match'
+import { ChevronLeftIcon } from '@lib/ui/icons/ChevronLeftIcon'
+import { VStack } from '@lib/ui/layout/Stack'
+import { useNavigateBack } from '@lib/ui/navigation/hooks/useNavigateBack'
+import { PageHeader } from '@lib/ui/page/PageHeader'
+import { PageHeaderIconButton } from '@lib/ui/page/PageHeaderIconButton'
+import { PageHeaderTitle } from '@lib/ui/page/PageHeaderTitle'
+import { useToast } from '@lib/ui/toast/ToastProvider'
+import { useCallback, useState } from 'react'
+import { useTranslation } from 'react-i18next'
+import { createGlobalStyle } from 'styled-components'
+
+const GlobalStyle = createGlobalStyle`
+  body {
+    background-image: url('/core/images/scanQRCodeBackground.png');
+    background-position: 0% 40%;
+    background-repeat: no-repeat;
+    background-size: cover;
+  }
+`
+
+export const UploadQrPage = () => {
+  const { t } = useTranslation()
+  const [{ title = t('keysign') }] = useCoreViewState<'uploadQr'>()
+  const [view, setView] = useState<'scan' | 'upload'>('scan')
+  const { addToast } = useToast()
+  const currentVaultId = useCurrentVaultId()
+  const vault = useVaults().find(vault => getVaultId(vault) === currentVaultId)
+  const deriveChainFromWalletAddress = useDeriveChainFromWalletAddress()
+  const navigate = useCoreNavigate()
+  const navigateBack = useNavigateBack()
+
+  const onScanSuccess = useCallback(
+    (value: string) => {
+      if (vault) {
+        const isURL = value.startsWith('http')
+
+        if (isURL) {
+          navigate({ id: 'deeplink', state: { url: value } })
+
+          return
+        }
+
+        const chain = deriveChainFromWalletAddress(value)
+        const coin = vault.coins.find(coin => coin.chain === chain)
+
+        if (coin) {
+          navigate({ id: 'send', state: { coin, address: value } })
+        } else {
+          addToast({ message: t('coin_not_found_in_current_vault') })
+        }
+      } else {
+        navigate({ id: 'deeplink', state: { url: value } })
+      }
+    },
+    [addToast, deriveChainFromWalletAddress, navigate, t, vault]
+  )
+
+  return (
+    <>
+      <VStack fullHeight>
+        <PageHeader
+          title={<PageHeaderTitle>{title}</PageHeaderTitle>}
+          primaryControls={
+            <PageHeaderIconButton
+              icon={<ChevronLeftIcon />}
+              onClick={() =>
+                view === 'scan' ? navigateBack() : setView('scan')
+              }
+            />
+          }
+          hasBorder
+        />
+        <Match
+          value={view}
+          scan={() => (
+            <ScanQrView
+              onUploadQrViewRequest={() => setView('upload')}
+              onFinish={onScanSuccess}
+            />
+          )}
+          upload={() => <UploadQrView />}
+        />
+      </VStack>
+      <GlobalStyle />
+    </>
+  )
+}
