+++ core/ui/qr/utils/decompressQrPayload.tsx
@@ -1,6 +1,8 @@
 import { getSevenZip } from '@core/mpc/compression/getSevenZip'
 
-export async function decompressQrPayload(value: string): Promise<Uint8Array> {
+export const decompressQrPayload = async (
+  value: string
+): Promise<Uint8Array> => {
   const bufferData = Buffer.from(value, 'base64')
   const sevenZip = await getSevenZip()
   sevenZip.FS.writeFile('data.xz', bufferData)
