--- lib/ui/qr/upload/UploadQRPage/UploadQRPage.styled.tsx
@@ -1,9 +1,0 @@
-import { PageContent } from '@lib/ui/page/PageContent'
-import styled from 'styled-components'
-
-export const StyledPageContent = styled(PageContent)`
-  background-image: url('/core/images/scanQRCodeBackground.png');
-  background-size: cover;
-  background-repeat: no-repeat;
-  background-position: 0% 40%;
-`
